Meer classes, sommige classes static gemaakt
