+++ InEenNotendop.Business/Note.cs
@@ -1,0 +1,40 @@
+﻿using NAudio.Midi;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace InEenNotendop.Business
+{
+
+    public class Note
+    {
+        public int NoteNumber { get; set; }
+        public TimeSpan NoteStartTime { get; set; }
+        public TimeSpan NoteDuration { get; set; }
+        public bool IsPlaying { get; set; }
+        public bool IsBlockGenerated { get; set; } = false;
+        public bool ScoreIsCalculated = false;
+
+        public Note(NoteOnEvent noteOnEvent, TimeSpan startTime) // voor midi input
+        {
+            NoteNumber = noteOnEvent.NoteNumber;
+            NoteStartTime = startTime;
+            IsPlaying = true;
+
+        }
+
+        public Note(int noteNumber, TimeSpan noteStartTime) // voor noot aanmaken zonder midi event
+        {
+            NoteNumber = noteNumber;
+            NoteStartTime = noteStartTime;
+        }
+
+        public void EndNote(TimeSpan endTime)
+        {
+            NoteDuration = endTime - NoteStartTime;
+            IsPlaying = false;
+        }
+    }
+}
